[ci] Increase timeout for CW310 Hyperdebug bitstream
The build time for FPGA bitstreams currently varies quite much, see
issue #17187.  PR #17216 addressed this problem for the CW310 bitstream
without Hyperdebug.

This commit increases the timeout for the CW310 Earl Grey *with
Hyperdebug* bitstream.  CI currently skips this check for PRs but runs
it after the merge, where it fails on `master` (again, see issue #17187
for further information).

Signed-off-by: Andreas Kurth <adk@lowrisc.org>