[hw,doc] Add one-line descriptions to all blocks
This commit adds one-line descriptions to all hardware blocks under
`hw/ip` that have a `data/<name>.hjson` file except `trial1`.

These one-line descriptions are intended for use in summary tables or as
subtitles.  As such, they should concisely convey the gist of a hardware
block.

For blocks that have their `.hjson` generated from a template
(`.hjson.tpl`), this commit updates the template and the generated
`.hjson`.  This commit also updates the `.hjson`s of the
top-level-specific blocks.

The new field that contains the one-line description, `one_line_desc`,
is added to `util/reggen` as optional.

Co-authored-by: Pirmin Vogel <vogelpi@lowrisc.org>
Signed-off-by: Andreas Kurth <adk@lowrisc.org>