[doc mv] contrib, get srt and prj gov for new doc
Moved files for the contributing and project government sections,
and the getting started book.

No file has been modified, so that git can more easily track moved files.

Exceptions to this are `{hw,ug}/_index.md` which had no content
and `hw_checklist.md.tpl` which is a symlink that needed updating.

Signed-off-by: Hugo McNally <hugo.mcnally@gmail.com>
Co-authored-by: Harry Callahan <hcallahan@lowrisc.org>
Co-authored-by: James Wainwright <james.wainwright@lowrisc.org>