[prim] Weighted Round Robin Arbiter
This commit implements an Weighted Round Robin (WRR) arbiter. This
arbiter chooses evenly among the received requests until all weights are
consumed. Another way for WRR is to consume a request's weight first
then move to the next requester.

This arbiter uses Parallel Prefix Computation to choose a winner and to
make the mask. BINTREE may be helpful as we've seen in
`prim_arbiter_tree` module.

The weight is given via `weight_i` port not from a parameter.

Signed-off-by: Eli Kim <eli@opentitan.org>