[hw,doc] Add one-paragraph descriptions to all blocks
This commit adds one-paragraph descriptions to all hardware blocks under
`hw/ip` that have a `data/<name>.hjson` file except `trial1`.

These one-paragraph descriptions are intended for use as high-level
summary and should describe the main elements and intentions of a
hardware block.

For blocks that have their `.hjson` generated from a template
(`.hjson.tpl`), this commit updates the template and the generated
`.hjson`.  This commit also updates the `.hjson`s of the
top-level-specific blocks.

The new field that contains the one-paragraph description,
`one_paragraph_desc`, is added to `util/reggen` as optional.

Co-authored-by: Pirmin Vogel <vogelpi@lowrisc.org>
Co-authored-by: Marno van der Maas <mvdmaas@lowrisc.org>
Co-authored-by: Harry Callahan <hcallahan@lowrisc.org>
Co-authored-by: Abdullah Varıcı <abdullah.varici@lowrisc.org>
Co-authored-by: Canberk Topal <ctopal@lowrisc.org>
Co-authored-by: Vladimir Rozic <vrozic@lowrisc.org>
Co-authored-by: Saad Khalid <saad.khalid@lowrisc.org>
Co-authored-by: Adrian Lees <a.lees@lowrisc.org>
Signed-off-by: Andreas Kurth <adk@lowrisc.org>