[opentitantool] Eeprom primitives in Spi trait
This PR introduces a new method on the spi::Target trait, for doing SPI
transactions following the EEPROM format, that is, consisting of a few
bytes of command/address, followed by a chuck of data to be either read
or written, optionally followed by repeated polling a status bit, to see
if the operation is complete.

A default implementation of the new method is provided for transport
backends which support generic SPI read/write transactions, in effect
moving part of the code from the existing SpiFlash struct into the
spi::Target trait.

Going forward, transports that do not implement generic SPI
transactions, but supports only EEPROM/Flash-style protocol, can provide
their own implementation of the new trait method, and leave the existing
generic Spi transaction method either unimplemented, or severely
restricted.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ibc33c37e7200eccdea2b2b5a370b4cd924c50bdd