[opentitantool] Atomic gpio manipulation on HyperDebug
Take advantage of recent addition to `Gpio` trait and to HyperDebug
firmware, to be able to simultaneously set opend drain / push pull, as
well as output level and weak pulling mode of a GPIO pin.  This avoids
glitches on the signals, which could otherwise happen if mode and level
were separately set.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ic27ccba95f1f30eb9315750ddc2bf29250da665c