[opentitantool] Propagate `transport init` to backend
The existing command `opentitantool transport init` will send a series
of instructions to the backend to configure pins according to a set of
configuration files.  Ideally, that would bring the transport into a
known state.  However, sometimes the configuration is under-specified,
that is, maybe it omits declaring the direction of a pin as "Input",
relying on that being the default for all pins at power-on.  Doing so
opens up the possibility that if the debugger device has previously been
used with a different configuration file, which used the same pin as
output, then after `transport init` the pin would remain an output, to
the surprise of automated tests or operators.

This PR makes `transport init` first tell the backend transport to reset
to power-on state, before sending the usual series of pin-specific
instructions, hopefully reducing the risk of surprises.

Some backends, such as UltraDebug, do not support re-configuraion of
pin direction, and others, such as Verilator, do not have backend
hardware which could carry over state from one test to the next, so a
default empty implementation of the new method is provided.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I0cee4a994ff72c653f0522fcfe92ce1be1ecf6af