[test] use correct OTP image in DV for power virus test
The default OTP image was being used for the power virus test which was
causing a test failure in DV. This fixes the DV sim config to specify
the correct OTP image.

Additionally, this temporarily disables flash scrambling / ECC to
silience DV testbench errors for the time being to unblock test
developement. These will be re-enabled in the future, and the relating
issues will be resolved.

Signed-off-by: Timothy Trippel <ttrippel@google.com>