[sw/vendor/riscv_compliance] Replace references to NexysVideo with CW310
We have stopped using and supporting the NexysVideo target already a
while ago. We haven't been running the RISC-V compliance suite on FPGA
for some time now but it's most likely to be run on the CW310 anyway
as only CW310 has full debug support (GDB + OpenOCD was used in the past
to launch RISC-V compliance tests).

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>