[opentitantool] End session when parent process terminates
The `opentitansession` can work as a daemon that connects to a backend
transport, and then detaches from the foreground terminal, as other unix
daemons.  The daemon can be told to stop by a signal.  This is useful at
times.

At other times, if one is writing a script for instance, which wants to
start the daemon at the top of the script, and keep it alive while the
script runs, it is not always convenient to ensure that if the script
terminates in abnormal ways, that it remembers to send a signal to the
`opentitansession` daemon.

This PR introduces a `--foreground` mode, which causes
`opentitansession` to not detach from the controlling terminal, and keep
running only as long as its parent process is alive.

Unfortunately, the Linux `prctl()` syscall is not part of the Rust `nix`
crate, so a bit of unsafe C linking of was necessary.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I742c11976582ce21b620e87868f5747fa3beba09