[wip] Change `unique case inside` to `unique case`
VCS cannot handle a range in `case inside` constructs when X propagation
instrumentation is enabled.  To avoid problems with this incompatibility,
this commit converts ranges to `if` clauses with two non-strict inequality
operators joined by a logical AND, moves them to the `default` section,
and removes `inside` from `unique case`.  The resulting code is equivalent
for synthesis purposes because the different branches remain of equal
priority to each another.  In simulation, however, the overlap checks that
come with `unique case` no longer include the ranges.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>