[aes, dv] Handle cases where message processing ends before FI
Depending on the number of messages and message length as well the
DUT peformance, the thread responsible for the message processing might
actually finish before any fault is injected into the DUT. In this case,
the message processing thread will not detect the alert, reset the DUT
and restart it.

This led to quite some tests timing out for the unmasked implementation
(running at higher performance). This commit solves the issue by letting
the thread who is injecting the fault wait for the alert and perform the
reset in case the message processing has finished before injecting the
fault.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>