[rom_ext] Add rom_ext verification key domain
This commit adds a new set of keys which are used by the ROM_EXT to
verify the first Silicon Owner boot stage. The keys are built into the
ROM_EXT binary and are covered by the code signature verified by the
ROM.

This commit does not support key invalidation via OTP. This is something
that will be added in a follow up change.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>