Add bitstream loading to hyperdebug
The hyperdebug board is not capable of loading a bitstream into the
CW310 board.  To accomplish bitstream loading, we add  a `CW310Flavor`
to the hyperdebug backend.  When using the CW310Flavor, we assume
that both the Hyperdebug and CW310 USB interfaces are attached to the
host.  The hyperdebug `load-bitstream` operation is delegated to the
CW310 connection.  All other IO (uarts, spi, etc) use the hyperdebug
connection.

Signed-off-by: Chris Frantz <cfrantz@google.com>