[otbn, rtl] Blank flag generation in bignum ALU for SCA hardening
Previously, there was a chance that transient effects between flags (in
particular MSB and LSB flags for logical operations) belonging to
different instructions could result in undesirable SCA leakage.

To prevent this, this commit blanks the data paths relevant for
managing and using the flags inside the bignum ALU similar to what is
done for the RF write data path. There are however two notable
exceptions where blanking cannot be applied without impacting the
performance of the architecture:
- Whether the Z flag for MAC operations is updated depends on the
  result of the MAC operation itself. This can result in SCA leakage
  related to previous Z flags. This commit slightly improves the
  situation by removing potential SCA leakage related to previous Z flag
  values of the other flag group.
- The BN.SEL selects either the data on read port A or B of the register
  file based on a single flag of one of the two flag groups. Even though
  flag selection is now blanked, there may be SCA leakage between the RF
  read ports as the flag selection result which serves as a selector
  input for the actual data is generated in the same pipeline stage.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>