[opentitantool] Add Dediprog backend
SPI support for Dediprog, it is able to program firmware into a CW310
and, and read from an EEPROM chip.

The voltage level of Dediprog is exposed as a GPIO pin "VCC", which is
permanently in AnalogOutput mode.

The driver suffers from not being able to read IO levels, voltage, or
SPI clock speed from the device, only set it.  So every time the driver
is initialized, it will reset to default values.

For the reasons above, the drivers initial VCC voltage is zero
volts (disabled).  The default `dediprog.json` declares 3.5 volt, which
gets applied by `transport init`.  Each command line invocation of
opentitantool will re-set the driver state, so all commands including
`transport init` must be given in a single invocation, e.g.:
```
opentitantool --interface=dediprog --exec 'transport init' bootstrap ...
```

Using `opentitansession` will mitigate the above issues, as the session
daemon will keep the driver open, and "remember" the values of GPIO pins
and current voltage/speed.  The same is the case for Rust programs that
directly use opentitanlib.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I275a514dff0086f3e1b3449a9bf2002bb8e3f00e