[aes, dv] Don't abort simulation when DUT is busy and requesting entropy
Previously, when the DUT was busy, but didn't set the stall flag, and
didn't have valid output, wasn't ready for new input the environment
assumed the DUT was stuck and would abort the simulation with a fatal
error after some time.

However, this condition is actually perfectly legal if the DUT is
requesting entropy for reseeding the PRNGs which may take a long time.
For this reason, this commit changes the check to only error out if no
entropy is requested.

This is related to lowRISC/OpenTitan#17399.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>