[util] refactor flash img processing script to enable scrambling
This refactors the flash image pre-processing script to enable
pre-scrambling a flash VMEM file so that it can be backdoor loaded in DV
sims when scrambling is enabled.

Extra command line arguments can now be passed to the script to turn on
the data scrambling feature. When scrambling is not enabled, the script
will simply reformat the input VMEM file to add integrity and
reliability ECC bits, same as before.

Signed-off-by: Timothy Trippel <ttrippel@google.com>