[sca/otbn] Move masking operations before capture
Currently some operations on the unmasked seed are performed inside
the batch capture loop. In other capture setups this situation was
causing problems with TVLA analysis due to data-dependent
residual charge on one of the board capacitors, creating a "fake
leakage" effect.
This commit moves the masking operation outside of the batch capture
loop, so that only masked data are used inside the loop.

Signed-off-by: Vladimir Rozic <vrozic@lowrisc.org>