[otbn, rtl] Improve blanking of flags, move clearing to predecoder
Previously, the `operation_commit_i` input was evaluated to determine
when to not update and/or clear the flags. However, as this signal
factors in many combinatorial error signals from the controller, the
signal is most likely very glitchy which is bad for blanking/SCA hard-
ening. Therefore this commit stops using this signal in the flag
generation/updating logic and migrates the corresponding functionality
to the instruction fetch/predecoder stage.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>