[opentitantool] Support SX1503 IO expanders
Development boards may emply IO expanders to control various features.
It is desirable to represent the signals of such IO expanders using the
`Gpio` trait, that is, with a name that can be used in `opentitantool`
like any "direct" GPIO pin of the debugger device.

This PR adds functionality to the `TransportWrapper` to translate the
`Gpio` operations into I2C transactions for the underlying transport
implementation, such that the IO lines of SX1503 IO expanders can be
manipulated.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I856701e44c0fe24ecb0c7448f4fb29de15a34201