[opentitantool] Introduce struct representing pin strapping
A Gpio trait implementation could represent a pin on an IO expander
sitting behind a MUXed I2C bus.  Gpio structs can outlive the
TransportWrapper object, and as such cannot call
`apply_pin_strapping()`.  This CL introduces an `PinStrapping` object,
such that `apply_pin_strapping()` is replaced with
`pin_strapping().apply()`.  This allows the Gpio implementation to hang
on to the `PinStrapping` object.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Icec1064dc020027ac39c56a79e38e970f24ec917