[ast] Bypass logic for clk_ext in FPGA targets
Bypass the dividers and derived clock generation for the external clock
(IOC6) when using the FPGA. Instead, reuse the clocks generated within
the FPGA, which bypassed the oscillator modeling of the AST. Add an
optional divider to the clk_io chain, to adjust the clock frequency for
when top_earlgrey's clkmgr expects a lower frequency input. Note that
for the FPGA, this means that clk_sys will remain at 10 MHz, even for
the "48 MHz" ext_clk case.

This connects clocks for tests that switch to the external clock on
IOC6. Previously, the external clock tree was not driven. Note this does
not connect IOC6 to the clock tree; it substitutes the main clock tree
for the external one, since those clocks are always available and
"calibrated".

Signed-off-by: Alexander Williams <awill@opentitan.org>