[crypto] Fix usage of `csrng_send_app_cmd` for EDN
Since 5cbddc36d85756fbe86f2a615717881292405c83 (PR #17350), the
`csrng_send_app_cmd` function is additionally being used to write EDN
registers that interface CSRNG in hardware.  In that use case, CSRNG's
SW status register cannot be used to check the status of a command, but
`csrng_send_app_cmd` unconditionally did so, which caused hangs.

This commit adds an argument, `check_completion`, to
`csrng_send_app_cmd` with which the checks can be disabled when the
function is called to write EDN registers.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>