[manuf] Start adding DUT personalization logic
This change adds configuration of the root secret in the SECRET2 OTP
partitition as well as the silicon creator seed stored in info flash.

The test currently has some workarounds in place to implement OTP and
flash_ctrl logic until we switch testutil libraries to use status_t.

There is also a ft.c functest target used to run the test case.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>