[entropy_src/dv] Create separate test for RNG with XHT rsp
This commit changes the default value of the knob that tells tests to
always provide the default XHT response to 1 (true).  The rationale is
that this is the same behavior as in the current top-level design, so we
want to make sure that all existing tests work in an environment in
which `entropy_src` is always given the default XHT response.  The only
test affected by this is `entropy_src_rng`.

In order to not completely remove any test that provides time-varying
XHT responses, this commit creates a new test,
`entropy_src_rng_with_xht_rsps`, in which the `xht_only_default_rsp`
knob is set to zero.  This test is equivalent to the `entropy_src_rng`
test before this commit.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>