[opentitantool] Emulator RESET pin
In order to reduce the need for test scripts to be aware of the
differences between physical development boards, and host emulation,
this CL introduces a special RESET pin into host emulation.
Manipulation of this pin is translated into stopping and starting the
emulator.

Next up, it should be supported to set the "external" drive of GPIO pins
while the emulator is off, and have the emulated code see the external
drive upon startup, e.g. for software strappings.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ib8217c4945dc664d45c87baec45bd66c5d7b9fa4