[site,doc] Update Gcloud build steps to use existing container
With the recent changes to the util/site/build-docs.sh script, we
now require additional dependencies for the build.
Move the cloud-build to use the existing container, which already
has these dependencies present.

The main difference is the need for the working directory (/workspace by
default) to be owned by the 'dev' user, and for the build to be run
as this user.
To achieve this, overwrite the ENTRYPOINT of the container image.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>