[opentitanlib] fix hyper310 tests
Since #17480 merged, the hyper310 tests have been failing with:
_"Error: Inconsistent configuration of transport interface hyper310 vs.
hyperdebug."_ This was due to the fact that the `--interface=hyper310`
parameter did not match the `interface` parameter in the config file.
This attempts to fix this error, and renable hyper310 test.

Note, they are not currently run in CI, hence the regression.

Signed-off-by: Timothy Trippel <ttrippel@google.com>