[opentitantool] Avoid no-op transport firmware upgrades
The HyperDebug driver already has code to determine the current firmware
version, and the version of a given binary image.  This PR adds logic to
skip the transport firmware upgrade (with a warning), in case the same
firmware is already running on HyperDebug.  (A --force flag allows
overriding, for use in case of e.g. manually compiled firmwares with
non-unique version numbers.)

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I7e3fbdb273f3ed86d094edf57c3a14867e18bbbf