[util] un-permute OTP image bits when processing
This un-permutes the OTP image bits when processing the OTP image in the
flash ECC/scrambling script, as the scrambling enablement option is read
directly out of the OTP image.

This was preventing the power virus test from running in the closed
source environment where bits in the OTP images are permuted.

Signed-off-by: Timothy Trippel <ttrippel@google.com>