[opentitanlib] refactor TAP strapping and add RISC-V TAP
The old configuration was quite messy: the TAP STRAPs default
level were specified in the CW310 specific files and also the
RESET strapping was touching (only!) one of the TAP STRAPs.
The new configuration specifies the default TAP STRAPs in
the generic opentitan file, the CW310 only specifies the pin
mapping. The RESET strapping does not touch the TAP STRAPS but
the default TAP STRAPS corresponds to the RISC-V TAP so the old
behaviour is preserved. Also add a strapping for the RISC-V TAP.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>