[opentitantool] add context to uart stale lock file error
In a shared environment (like fred), if a user runs opentitantool
and grabs the UART, but for some reason does not stop properly,
it might leave behind a stale file that other users cannot remove.
This is really not obvious so this commit adds a bit of context
to the error so that at least the user knows what the problem is.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>