[dv/keymgr] Update testbench to not change OTP key on the fly
Fix issue #17640. As Fatih analynized, when compute kmac keys, keymgr
only samples the OTP key output once. But the timing is hard to predict
on the scb side.
The current scb samples the otp keys when the request happened, but if
the otp keys are changing on the fly, it might be sampled at a wrong
timestamp.
To avoid this failure, we now only randomize OTP keys once and keep it
stable. To test keymgr won't latch the wrong key when OTP keys are
changing on the fly, we plan to have a V3 direct sequence test for it.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>