[dv/otp] Fix otp_parallel_lc_esc seq
PR #17753 tries to fix a close source failure, but created a coverage
hole.
The intention of the sequence is to issue lc_esc and check afterwards
the OTP_CTRL is indeed locked up.
The fix from #17753 is that - when we call a new sequence, it always
starts with pre_start() and will call dut_init.

I think a cleaner solution is instead to use the knob `do_dut_init` to
avoid triggering reset.

So this PR:
1). Added a `do_dut_init` check before issuing dut_init in
  otp_ctrl_smoke_seq.
2). Simply the otp_ctrl_parallel_lc_esc logic to set `do_dut_init = 0`.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>