[doc] Slightly tweak getting started intro
The initial sentence could sound a bit like "you need to not ...
simulate the hardware it runs on", which is a bit confusing.

Also, reformat things about the best way to build stuff. The "however"
was possibly a little confusing, and splitting things before "if you
are new to the project" seems to make it a bit easier to read.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>