[rom/e2e] fix sram_program_fpga_cw310_test_otp_* tests
The issue is described in the code. In short, GDB tries to read from the stack when we set $sp
but the stack in the RAM which is not initialized. Since the RAM is scrambled, read from an
uninitialised location will trigger an ECC error and the read will fail. Due to a RTL bug, once
a memory access fails, all subsequence memory access through the debug module fail. We work around
this problem by initializing the RAM prior to running the code.
This ccommit also documents and changes the ePMP setup to only add the main RAM and not the entire
space in the PMP entry.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>