[gdb_test_coordinator] flush the output of all processes before quitting
At the moment, the script will either terminate openocd or just exit early
when a failure is detected. This is inconvenient because some important
output/error messages maybe still be in the pipe or some other process is
not dead yet and still printing. This commit gives a chance to print more
messages before exiting. Currently the timeout is one second but it can
easily be changed.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>